+++ pywatts/utils/_xarray_time_series_utils.py
@@ -1,11 +1,11 @@
-from typing import List, Dict
+from typing import List, Dict, Union
 
 import numpy as np
 import pandas as pd
 import xarray as xr
 
 
-def _get_time_indexes(x: Dict[str, xr.DataArray]) -> List[str]:
+def _get_time_indexes(x: Union[xr.DataArray, Dict[str, xr.DataArray]]) -> List[str]:
     indexes = []
     if isinstance(x, xr.DataArray):
         for k, v in x.indexes.items():